+++ aspnet-core/src/TicketTracker.Core/Entities/Subscription.cs
@@ -7,11 +7,11 @@ using System.Threading.Tasks;
 using TicketTracker.Authorization.Users;
 
 namespace TicketTracker.Entities {
-    public class Subscription : FullAuditedEntity<long, User> {
+    public class Subscription : FullAuditedEntity<int, User> {
         public User User { get; set; }
         public long UserId { get; set; }
 
         public Ticket Ticket { get; set; }
-        public long TicketId { get; set; }
+        public int TicketId { get; set; }
     }
 }
