+++ aspnet-core/src/TicketTracker.Core/Entities/Work.cs
@@ -7,12 +7,12 @@ using System.Threading.Tasks;
 using TicketTracker.Authorization.Users;
 
 namespace TicketTracker.Entities {
-    public class Work : FullAuditedEntity<long, User> {
+    public class Work : FullAuditedEntity<int, User> {
         public ProjectUser ProjectUser { get; set; }
-        public long ProjectUserId { get; set; }
+        public int ProjectUserId { get; set; }
 
         public Ticket Ticket { get; set; }
-        public long TicketId { get; set; }
+        public int TicketId { get; set; }
 
         public ushort? WorkedTime { get; set; }
         public ushort? EstimatedTime { get; set; }
